requirements.txt
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
pypdf
sentence-transformers
python-multipart
-e .